fix(pr459): apply Codex P1/P2 fixes and resolve CI errors/warnings
Codex review fixes:
- VulkanDevice::Shutdown now calls vkDeviceWaitIdle before
  TransientBufferAllocator::Shutdown so in-flight buffer handles are
  not destroyed while the GPU still references them (P1).
- ShaderDiskCache::HashSourceForDisk now mixes source.entryPoint into
  its FNV-1a chain so shaders compiled from the same HLSL file with
  different entry points no longer collide on the disk cache key (P2).

CI errors / warnings:
- VersionControlEnums.h: rename BranchOperation::{CREATE,DELETE,RENAME,
  MERGE,REBASE,SWITCH} to *_BRANCH variants and add defensive
  #undef DELETE/IN/OUT/OPTIONAL/ERROR at the top of the header so the
  Windows SDK macros (winnt.h) do not corrupt the enum body. Exposed
  by Phase AA wiring VersionControlSystem into the editor panel
  factory on Windows builds.
- TestLevelStreamingSystemPhaseAA.cpp: WorldTile contains
  std::future<bool> loadingTask which deletes the copy constructor;
  the tests now std::move(tile) into AddTile as required.
- SelectionManager.h: SPARK_LOG_DEBUG for entity selection used %u
  with a uint64_t EntityId. Now formats with %llu and an explicit
  cast to unsigned long long for portable printing.
- TestEntityPresetManagerPhaseEE.cpp, TestLODGeneratorPhaseGG.cpp,
  TestLevelStreamingSystemPhaseAA.cpp: remove unsigned `>= 0`
  tautologies that tripped -Wtype-limits; the assertions now just
  verify the call returned without crashing.

Local validation: linux-gcc-release build clean; ctest 100% passed.

+++ SparkEditor/Source/Panels/SelectionManager.h
@@ -203,7 +203,11 @@ namespace SparkEditor
             if (IsLocked(entityId))
                 return;
 
-            SPARK_LOG_DEBUG(Spark::LogCategory::Editor, "SelectionManager: Select entity %u", entityId);
+            // EntityId is a 64-bit unsigned type; format with %llu and cast
+            // to unsigned long long for portable printing (GCC/MSVC both
+            // warned about %u vs uint64_t on CI).
+            SPARK_LOG_DEBUG(Spark::LogCategory::Editor, "SelectionManager: Select entity %llu",
+                            static_cast<unsigned long long>(entityId));
             auto oldSelection = GetSelectionVector();
             m_selection.clear();
             m_selectionOrder.clear();
